% Generated by build script
\name{pypi_package}
\alias{pypi_package}
\title{Get PyPI package metadata}
\usage{
pypi_package()
}
\arguments{
\item{name}{Package name (e.g. "requests", "numpy", "pandas")}
}
\value{tibble: name, version, summary, author, author_email, license,}
\description{Get PyPI package metadata}
