% Generated by build script
\name{pypi_releases}
\alias{pypi_releases}
\title{Get PyPI package release history}
\usage{
pypi_releases()
}
\arguments{
\item{name}{Package name (e.g. "requests", "numpy")}
\item{limit}{Maximum number of versions to return (default 20)}
}
\value{tibble: version, upload_time, python_version, size,}
\description{Get PyPI package release history}
